pyproject.toml
[project]
name = "stain"
version = "0.1.0"
description = "Multi-agent text analysis pipeline that surfaces LLM generation patterns"
authors = [{ name = "Kai Hallett", email = "kai@oceanheart.ai" }]
readme = {text = "Multi-agent text analysis pipeline that surfaces LLM generation patterns", content-type = "text/plain"}
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
stain = "stain.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stain"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
